requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.20
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
redis>=5.0.0
pinecone-client>=3.0.0
chromadb>=0.4.22
pypdf>=4.0.0
python-docx>=1.1.0
tiktoken>=0.5.2
openai>=1.10.0
anthropic>=0.18.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
